refactor: explicitly reference tokenCount via a local variable
within `claimRewards` there are enough local variables that the
solidity optimizer does not implicitly cache `tokenCount` on
the stack. this means each iteration of the for loops requires
an SLOAD operation. by explicitly copying `tokenCount` to a local
variable we can force it to the stack and reduce gas costs.